backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
opencv-python==4.8.1.78
mediapipe==0.10.21  # Changed to available version
numpy==1.24.3
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0